first time all green :) - some tests only enabled for jvm - some more work needed
