+++ core/shared/src/main/scala/org/qirx/littlespec/Specification.scala
@@ -1,3 +1,7 @@
 package org.qirx.littlespec
 
+import scala.scalajs.js.annotation.{JSExportDescendentClasses, JSExportDescendentObjects}
+
+@JSExportDescendentObjects
+@JSExportDescendentClasses
 abstract class Specification extends BaseSpecification
