+++ core/shared/src/main/scala/org/qirx/littlespec/sbt/Task.scala
@@ -6,18 +6,13 @@ import org.qirx.littlespec.Specification
 import sbt.testing.{Task => BaseTask, _}
 
 case class Task[T <: Specification](
-  testClass: Class[T],
-  isObject: Boolean,
+  testInstance: Specification,
   taskDef: sbt.testing.TaskDef,
   reporter: SbtReporter) extends sbt.testing.Task {
 
   val tags: Array[String] = Array.empty
 
   def execute(eventHandler: EventHandler, loggers: Array[Logger]): Array[sbt.testing.Task] = {
-
-    val testInstance =
-      if (isObject) testClass.getField("MODULE$").get(null).asInstanceOf[T]
-      else testClass.newInstance
 
     val results = testInstance.executeFragments()
