+++ core/shared/src/test/scala/org/qirx/littlespec/SpecificationSpec.scala
@@ -8,13 +8,14 @@ import org.qirx.littlespec.assertion.ThrowingAssertions
 import org.qirx.littlespec.fragments.ExampleFragments
 import org.qirx.littlespec.fragments.FragmentContainer
 import org.qirx.littlespec.fragments.Fragment
+import testUtils.NoJSExportSpecification
 
 object SpecificationSpec extends Specification {
 
   "Specification should" - {
 
     "be of the correct type" - {
-      val specification = new Specification {}
+      val specification = new NoJSExportSpecification {}
       type Fragments = FragmentContainer with ExampleFragments
       type DefaultAssertions = StaticAssertions with ThrowingAssertions with TypeAssertions
       type DefaultAssertEnhancements = BasicAssertEnhancements with FragmentBodyAssertEnhancement
@@ -23,7 +24,7 @@ object SpecificationSpec extends Specification {
     }
 
     "have a type alias for body" - {
-      val specification = new Specification {}
+      val specification = new NoJSExportSpecification {}
       val x: specification.FragmentBody = todo: Fragment.Body
       success
     }
